fix: Case-insensitive breakpoint matching, unused var, overflow type
- Use strcasecmp for breakpoint path matching — UserTalk identifiers
  are case-insensitive, so "mainResponder.respond" must match
  "mainresponder.respond".
- Suppress unused hroot warning with (void)hroot.
- Change script_stack_overflow from short to int for consistency.

+++ frontier-cli/debug_handler.h
@@ -103,7 +103,7 @@ typedef struct tydebugstate {
     char current_script[DEBUG_SCRIPT_PATH_MAX]; /* current script dotted path */
     char script_stack[DEBUG_SCRIPT_STACK_MAX][DEBUG_SCRIPT_PATH_MAX]; /* saved caller paths */
     short script_stack_depth;                   /* stack pointer (0 = empty) */
-    short script_stack_overflow;                /* push/pop balance when stack overflows */
+    int script_stack_overflow;                  /* push/pop balance when stack overflows */
 } tydebugstate, *ptrdebugstate;
 
 /*
